Add a Time module
Isolate START_TIME_TYPE definition in a header. This allows the
START_TIME_TYPE to be used as a unique time type definition. And
to add common functions prototypes despite the conditional
compilation.

Centralize time functions. Which prevents coding errors on this
functions and reduces library's .text section size.

Notes:
1) The start_clock() had a static variable, start_time. Since
   this variable is shared by all callers, it could generate
   concurrency problems like data races. Thus, this function
   is now reentrant.
2) The Paho library exports only symbols from MQTTAsync and
   MQTTClient. Thus, the modules Time, Thread, StackTrace and
   Log had to be linked against the tests.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>